Sécurité pour éviter de supprimer la base

+++ project/bin/delete_from_object.sh
@@ -1,5 +1,8 @@
 . bin/config.inc
 
-	
 OBJREV=$(curl -s $1 | sed 's/{"_id":"//' | sed 's/","_rev":"/?rev=/' | sed 's/".*//')
-curl -s -X DELETE http://$COUCHHOST:$COUCHPORT/$COUCHBASE/$OBJREV
+if test "$OBJREV" ; then
+	curl -s -X DELETE http://$COUCHHOST:$COUCHPORT/$COUCHBASE/$OBJREV
+else
+	echo "Document non trouvé. Il faut mettre l'url du document"
+fi
